+++ packages/cli/CHANGELOG.md
@@ -7,6 +7,8 @@
 - Added an `extends` setting to `.moon/workspace.yml` and `.moon/project.yml`.
 - Added a `actionRunner.logRunningCommand` setting to `.moon/workspace.yml` for logging the task
   command being ran.
+- Added a global `--logFile` option to the CLI. Also supports a new `MOON_LOG_FILE` environment
+  variable.
 
 #### 🐞 Fixes